face_app/requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==3.1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
hiredis==1.1.0
html5lib==0.9999999
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
ml2rt==0.2.0
numpy==1.19.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
psutil==5.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
redis==3.5.3
redisai==1.0.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.25.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
